osc/fsm: Align everything to cache line
This is done reduce the possibility of false sharing as the GenZ
hardware only allows writes and reads on cache line level. Also there is
no guarantee of cache coherency across nodes. Therefore this commit is
not yet final as it only covers alignment.

In result of covering alignments the datatypes had to be changed thus I
tried to check use 64bit datatypes if possible and 32bit if only 32 bit
atomics are supported.

Signed-off-by: guserav <erik.zeiske@hpe.com>