osc/fsm: Store more information about maped region
For atomics we need more information about the region:
remote_keys
remote_vaddr

As the remote_vaddr can be zero based instead of the actual vaddr on the
remote side we also have to check how the provider was initialised.
Therefore this commit also stores the provider struct in the common
component so that it can be checked for the FI_MR_VIRT_ADDR value set.

Signed-off-by: guserav <erik.zeiske@hpe.com>