osc/fsm: Add distinction between flush and commit
It might be important to have a distinction between flush and commit for
later. This would be needed if it is not necessary to wait on the
completion of commit. As flush and invalidate are currently doing the
same thing under the cover it would have been possible to just use
invalidate when we are trying to flush but in the future it might be
possible to have an invalidate that really is not flushing.

Signed-off-by: guserav <erik.zeiske@hpe.com>