osc/fsm: Fix passive target synchronisation
Add remarks about the atomics that need to be waited on.

No need to invalidate all window copies if locking a specific target.

Even with shared access it is possible to issue puts on the window so we
need to flush it anyway. This is especially needed since nocheck could
be both shared and exclusive.

Dropping unnecessary memory barriers that are already part of the commit
and invalidate calls.

Signed-off-by: guserav <erik.zeiske@hpe.com>